requirements.txt
spacy==3.7.2
pymupdf==1.24.9
python-docx==1.1.2
reportlab==4.2.2
transformers==4.52.1
torch==2.3.0
faiss-cpu==1.8.0
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
beautifulsoup4==4.13.4
nltk
rapidfuzz
scikit-learn
joblib
PyMuPDF
spacy
python-docx
docx2pdf

